requirements.txt
asn1crypto>=0.24.0
botbuilder-schema>=4.0.0a0
botframework-connector>=4.0.0a0
certifi>=2018.1.18
cffi>=1.11.5
chardet>=3.0.4
cognitive-luis>=0.1
cryptography>=2.2.2
idna>=2.6
isodate>=0.6.0
msrest>=0.4.27
oauthlib>=2.0.7
pip>=9.0.3
pycparser>=2.18
PyJWT>=1.6.1
requests>=2.18.4
requests-oauthlib>=0.8.0
setuptools>=28.8.0
six>=1.11.0
urllib3>=1.22
